+++ NearFuture/SettingsView.swift
@@ -140,168 +140,9 @@ struct SettingsView: View {
 	}
 }
 
-struct iCloudSettingsView: View {
-	@State var viewModel: EventViewModel
-	@State var showPushAlert: Bool = false
-	@State var showPullAlert: Bool = false
-	
-	@Binding var hasUbiquitous: Bool
-	@Binding var lastSyncWasSuccessful: Bool
-	@Binding var lastSyncWasNormalAgo: Bool
-	@Binding var localCountEqualToiCloud: Bool
-	@Binding var icloudCountEqualToLocal: Bool
-	
-	var updateStatus: () -> Void
-	
-	var body: some View {
-		List {
-			HStack {
-				Spacer()
-				VStack {
-					ZStack {
-						Image(systemName: "icloud")
-							.resizable()
-							.scaledToFit()
-							.frame(width: 75, height: 75)
-							.symbolRenderingMode(.multicolor)
-						Text("\(viewModel.icloudEventCount)")
-							.font(.title2)
-					}
-					HStack {
-						Button(role: .destructive) {
-							showPushAlert.toggle()
-						} label: {
-							Image(systemName: "arrow.up")
-								.resizable()
-								.scaledToFit()
-								.frame(width: 30, height: 40)
-						}
-						.buttonStyle(BorderedButtonStyle())
-						.alert("Warning", isPresented: $showPushAlert) {
-							Button("OK", role: .destructive) {
-								viewModel.replaceiCloudWithLocalData()
-								viewModel.sync()
-								updateStatus()
-							}
-							Button("Cancel", role: .cancel) {}
-						} message: {
-							Text("This will replace Events stored in iCloud with Events stored locally.")
-						}
-
-						Button() {
-							viewModel.sync()
-							updateStatus()
-						} label: {
-							Image(systemName: "arrow.triangle.2.circlepath")
-								.resizable()
-								.scaledToFit()
-								.frame(width: 30, height: 40)
-								.foregroundStyle(Color.accentColor)
-						}
-						.buttonStyle(BorderedButtonStyle())
-						
-						Button(role: .destructive) {
-							showPullAlert.toggle()
-						} label: {
-							Image(systemName: "arrow.down")
-								.resizable()
-								.scaledToFit()
-								.frame(width: 30, height: 40)
-						}
-						.buttonStyle(BorderedButtonStyle())
-						.alert("Warning", isPresented: $showPullAlert) {
-							Button("OK", role: .destructive) {
-								viewModel.replaceLocalWithiCloudData()
-								viewModel.sync()
-								updateStatus()
-							}
-							Button("Cancel", role: .cancel) {}
-						} message: {
-							Text("This will replace Events stored locally with Events stored in iCloud.")
-						}
-					}
-					ZStack {
-						Image(systemName: "iphone.gen3")
-							.resizable()
-							.scaledToFit()
-							.frame(width: 75, height: 75)
-							.symbolRenderingMode(.monochrome)
-						Text("\(viewModel.localEventCount)")
-							.font(.headline)
-					}
-				}
-				Spacer()
-			}
-			.onAppear {
-				viewModel.sync()
-				updateStatus()
-			}
-			
-			HStack {
-				Circle()
-					.frame(width: 20, height: 20)
-					.foregroundStyle(hasUbiquitous ? .green : .red)
-				Text("iCloud Key Value Store:")
-				Text("\(hasUbiquitous ? "" : "Not ")Working")
-					.bold()
-			}
-			
-			HStack {
-				Circle()
-					.frame(width: 20, height: 20)
-					.foregroundStyle(lastSyncWasSuccessful ? .green : .red)
-				Text("Sync Status:")
-				Text("\(viewModel.syncStatus)")
-					.bold()
-			}
-			
-			HStack {
-				Circle()
-					.frame(width: 20, height: 20)
-					.foregroundStyle(lastSyncWasNormalAgo ? .green : .red)
-				Text("Last Sync:")
-				Text("\(viewModel.lastSync?.formatted() ?? "Never")")
-					.bold()
-			}
-			
-			HStack {
-				Circle()
-					.frame(width: 20, height: 20)
-					.foregroundStyle(localCountEqualToiCloud ? .green : .red)
-				Text("\(viewModel.localEventCount)")
-					.bold()
-				Text("Local Events")
-			}
-			
-			HStack {
-				Circle()
-					.frame(width: 20, height: 20)
-					.foregroundStyle(icloudCountEqualToLocal ? .green : .red)
-				Text("\(viewModel.icloudEventCount)")
-					.bold()
-				Text("Events in iCloud")
-			}
-		}
-		.navigationTitle("iCloud")
-		.navigationBarTitleDisplayMode(.inline)
-	}
-}
-
 #Preview {
 	SettingsView(
 		viewModel: EventViewModel()
-	)
-}
-
-#Preview("iCloudSettingsView") {
-	iCloudSettingsView(
-		viewModel: EventViewModel(),
-		hasUbiquitous: .constant(true),
-		lastSyncWasSuccessful: .constant(true),
-		lastSyncWasNormalAgo: .constant(true),
-		localCountEqualToiCloud: .constant(true),
-		icloudCountEqualToLocal: .constant(true),
-		updateStatus: test
 	)
 }
